+++ src/point.c
@@ -12,12 +12,12 @@
 // Defining the point structure:
 struct point
 {
-    int x, y; // Cartesian coordinates of the point
+    float x, y; // Cartesian coordinates of the point
     int group;  // group the point belongs
 };
 
 // Defining the function that creates a new point:
-Point *Point_Create(int num, int x, int y)
+Point *Point_Create(int num, float x, float y)
 {
     Point *p = malloc(sizeof(*p));
     p->x = x;
@@ -67,13 +67,13 @@ int Point_GetGroup(Point *point)
 }
 
 // Defining the function that returns the x coordinate of the point:
-int Point_GetX(Point *point)
+float Point_GetX(Point *point)
 {
     return point->x;
 }
 
 // Defining the function that returns the y coordinate of the point:
-int Point_GetY(Point *point)
+float Point_GetY(Point *point)
 {
     return point->y;
 }
